+++ common/src/client/java/io/github/jamalam360/jamlib/client/impl/JamLibClient.java
@@ -2,9 +2,11 @@ package io.github.jamalam360.jamlib.client.impl;
 
 import static io.github.jamalam360.jamlib.JamLib.JAR_RENAMING_CHECKER;
 
+import io.github.jamalam360.jamlib.api.network.Network;
 import io.github.jamalam360.jamlib.client.api.events.ClientMouseScrollEvents;
 import io.github.jamalam360.jamlib.client.api.events.ClientConnectionEvents;
-import io.github.jamalam360.jamlib.impl.network.NetworkCapabilitiesImpl;
+import io.github.jamalam360.jamlib.client.impl.network.ClientNetworkCapabilitiesImpl;
+import io.github.jamalam360.jamlib.impl.network.CapabilitiesPacket;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.LocalPlayer;
@@ -15,10 +17,11 @@ public class JamLibClient {
     @ApiStatus.Internal
     public static void init() {
         ClientConnectionEvents.CONNECT.listen(JamLibClient::onPlayerJoin);
-        ClientConnectionEvents.CONNECT.listen(_ -> NetworkCapabilitiesImpl.onClientJoin());
 
         ClientMouseScrollEvents.IN_SCREENS.listen((m, a) -> ClientMouseScrollEvents.ALWAYS.invokeCancellable(l -> l.onScroll(m, a)));
         ClientMouseScrollEvents.OUT_OF_SCREENS.listen((m, a) -> ClientMouseScrollEvents.ALWAYS.invokeCancellable(l -> l.onScroll(m, a)));
+
+        Network.registerHandler(Network.Direction.CLIENT_BOUND, CapabilitiesPacket.TYPE, ClientNetworkCapabilitiesImpl::handleServerCapabilities);
     }
 
     private static void onPlayerJoin(Minecraft minecraft) {