+++ common/src/main/java/io/github/jamalam360/jamlib/impl/network/JamLibPacket.java
@@ -3,7 +3,6 @@ package io.github.jamalam360.jamlib.impl.network;
 import io.github.jamalam360.jamlib.JamLib;
 import io.github.jamalam360.jamlib.api.network.PayloadType;
 import net.minecraft.network.RegistryFriendlyByteBuf;
-import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.resources.Identifier;